* Fix [conf_ver2.c] Clean up the log callback when the configuration file parsing fails
Normally, nanomq starts with a log_callback with log_level warning and
executes log_clear_callback after parsing the configuration file. Clean
up the callback with log_level warning and add a new log_callback.

But when a log_level parameter is passed(like --log_level=trace) and the
configuration file is not found, it does not clean up the log_callback,
which leads to duplicate log_callbacks, and eventually duplicate
printing problems.