URL u_port should be a number not a string.
The idea here is to reduce the dynamic allocations used for
URLs, and also the back and forth with parsing begin strings
and port numbers.  We always resolve to a port number, and
this is easier for everyone.

The real goal in the long term is to eliminate dynamic allocation
of the URL fields altogether, but that requires a little more
work.  This is a step in the right direction.

Signed-off-by: jaylin <jaylin@emqx.io>